Added playlist for specific emotions
