+++ Moodis/Moodis/Feature/MP3Player/MusicPlayer.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Music
+{
+    class MusicPlayer
+    {
+        [DllImport("winmm.dll")]
+        private static extern long mciSendString(string lpstrCommand, StringBuilder lpstrReturnString, int uReturnLength, int hwdCallBack);
+
+        public void open(string File)
+        {
+            string Format = @"open ""{0}"" type MPEGVideo alias MediaFile";
+            string command = string.Format(Format, File);
+            mciSendString(command, null, 0, 0);
+        }
+
+        public void play()
+        {
+            string command = "play MediaFile";
+            mciSendString(command, null, 0, 0);
+        }
+        public void stop()
+        {
+            string command = "stop MediaFile";
+            mciSendString(command, null, 0, 0);
+        }
+    }
+}
