+++ Moodis/Moodis/Moodis.csproj
@@ -104,6 +104,7 @@
     <Compile Include="Feature\Login\SignIn\SignInForm.Designer.cs">
       <DependentUpon>SignInForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="Feature\MP3Player\MusicPlayer.cs" />
     <Compile Include="Feature\Statistics\CalendarForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -190,7 +191,7 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Feature\NewFolder1\" />
+    <Folder Include="Feature\MP3Player\Music\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
